* polla/master:
  Use _localtime32() instead of localtime() on Windows
  Do not misuse long to cast a pointer
  Win32 support
  Implement is_full_path() helper function

+++ snprintf.c
@@ -342,7 +342,7 @@ static size_t dopr(char *buffer, size_t maxlen, const char *format, va_list args
 				break;
 			case 'p':
 				strvalue = (char *)va_arg(args, void *);
-				fmtint (buffer, &currlen, maxlen, (long) strvalue, 16, min, max, flags);
+				fmtint (buffer, &currlen, maxlen, (intptr_t) strvalue, 16, min, max, flags);
 				break;
 			case 'n':
 				if (cflags == DP_C_SHORT) {
